requirements.txt
flask==2.0.1
opencv-python-headless==4.8.0.74
numpy==1.24.3
Pillow==9.5.0
flask-cors==3.0.10
werkzeug==2.0.3 